+++ news-homepage/src/components/molecules/NewSection.tsx
@@ -2,10 +2,19 @@ import { newsContents } from "../../config/contents";
 import { Heading } from "../atoms/Heading";
 import { NewsItem } from "./NewsItem";
 
-export const NewSection = () => {
+interface NewSectionProps {
+  id: string;
+}
+
+export const NewSection = ({ id }: NewSectionProps) => {
   return (
-    <div className="my-14 w-full space-y-6 bg-very-dark-blue p-4 py-6">
-      <Heading className="text-3xl font-semibold text-soft-orange">New</Heading>
+    <div className="my-14 w-full space-y-6 bg-very-dark-blue p-4 py-6 xl:my-0 xl:p-6 xl:py-8">
+      <Heading
+        id={id}
+        className="text-3xl font-semibold text-soft-orange xl:text-4.5xl"
+      >
+        New
+      </Heading>
 
       <div>
         {newsContents.map((news, index) => (
@@ -14,6 +23,9 @@ export const NewSection = () => {
             heading={news.title}
             paragraph={news.content}
             isLast={index === newsContents.length - 1}
+            containerClass="xl:my-8"
+            headingClass='xl:text-xl lg:hover:text-soft-orange'
+            paragraphClass='xl:text-sm'
           />
         ))}
       </div>